test: add curated AI-generated unit tests
